Renamed and refactored packages, entities, services, configs, etc...
