--- src/main/java/com/compilercharisma/chameleonbusinessstudio/appointments/package-info.java
@@ -1,16 +1,0 @@
-/**
- * this package is responsible for managing the appointment data used by the
- * program.
- * +---------------------------+-----------------------------------------------+
- * |         Class             |                  Purpose                      |
- * +---------------------------+-----------------------------------------------+
- * | AppointmentController     | handles HTTP reqs to 'appointments' endpoint  |
- * | AppointmentEntity         | data object                                   |
- * | AppointmentModelAssembler | adds page links to pagination resources       |
- * | AppointmentRepository     | data access object                            |
- * | AppointmentService        | business logic for appointments. Use this!    |
- * | AppointmentSpecifications | used to filter in database queries            |
- * | AppointmentTagEntity      | part of data object                           |
- * +---------------------------+-----------------------------------------------+
- */
-package com.compilercharisma.chameleonbusinessstudio.appointments;
