--- src/main/java/com/compilercharisma/chameleonbusinessstudio/authentication/NoUserLoggedInException.java
@@ -1,13 +1,0 @@
-package com.compilercharisma.chameleonbusinessstudio.authentication;
-
-/**
- *
- * @author Matt Crow <mattcrow19@gmail.com>
- */
-public class NoUserLoggedInException extends RuntimeException {
-    public NoUserLoggedInException() {}
-
-    public NoUserLoggedInException(String msg) {
-        super(msg);
-    }
-}
