--- src/main/java/com/compilercharisma/chameleonbusinessstudio/authentication/UserNotRegisteredException.java
@@ -1,14 +1,0 @@
-package com.compilercharisma.chameleonbusinessstudio.authentication;
-
-/**
- *
- * @author Matt Crow <mattcrow19@gmail.com>
- */
-public class UserNotRegisteredException extends RuntimeException {
-    public UserNotRegisteredException() {
-    }
-
-    public UserNotRegisteredException(String msg) {
-        super(msg);
-    }
-}
