+++ src/main/java/com/compilercharisma/chameleonbusinessstudio/repository/AppointmentRepository.java
@@ -1,0 +1,18 @@
+package com.compilercharisma.chameleonbusinessstudio.repository;
+
+import java.time.LocalDateTime;
+import java.util.List;
+
+import com.compilercharisma.chameleonbusinessstudio.entity.AppointmentEntity;
+import org.springframework.data.domain.Pageable;
+import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
+import org.springframework.data.repository.PagingAndSortingRepository;
+
+/** 
+ * can use findAll(AppointmentSpecification.xyz)
+ * @author Matt Crow <mattcrow19@gmail.com>
+ */
+public interface AppointmentRepository extends PagingAndSortingRepository<AppointmentEntity, Integer>, JpaSpecificationExecutor<AppointmentEntity> {
+    public List<AppointmentEntity> findAllByStartTimeGreaterThanAndEndTimeLessThan(LocalDateTime start, LocalDateTime end);
+    public List<AppointmentEntity> findAllByStartTimeGreaterThanAndEndTimeLessThan(LocalDateTime start, LocalDateTime end, Pageable sort);
+}
